% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homoscedasticity.R
\name{homoscedasticity}
\alias{homoscedasticity}
\title{homoscedasticity}
\usage{
homoscedasticity(data, formula, threshold = 0.05)
}
\arguments{
\item{data}{a dataframe containing regression data}

\item{formula}{a formula in the format "y ~ x1 + x2 + ..." indicating regression variables}

\item{threshold}{a float indicating rejection threshold of correlation coefficient}
}
\value{
a plot of predicted values vs residuals an the correlation p value
}
\description{
This function recieves a linear regression model and outputs a
scatter plot figure of residuals plotted against fitted values. It prints
a statement indicating the results of the homoscedasticity test and outputs
a dataframe containing spearman correlation coefficients between the
absolute residuals and the fitted y values.
}
\examples{
library(tibble)
data <- tibble(X = c(1,2,3,4,5), y=c(1,2,4,4,6))
formula <- "y ~ X"
homoscedasticity(data, formula)

}
