% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicollinearity.R
\name{multicollinearity}
\alias{multicollinearity}
\title{multicollinearity}
\usage{
multicollinearity(X, vif_threshold = 10)
}
\arguments{
\item{X}{a dataframe containing predictor data}

\item{threshold}{a float indicating threshold for VIF}
}
\value{
a dataframe containing VIF coefficients and a statement whether the assumption is violated.
}
\description{
his function recieves a data frame of predictors and outputs VIF for each feature
along with a stetement whether or not the multicollinearity assumption is violated.
}
\examples{
library(tibble)
X <- tibble(c(1,2,3,4,5))
multicollinearity(X, 10)

}
