% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality.R
\name{normality}
\alias{normality}
\title{normality}
\usage{
normality(data, formula, p_threshold = 0.05)
}
\arguments{
\item{data}{a dataframe containing regression data}

\item{formula}{a formula in the format "y ~ x1 + x2 + ..." indicating regression variables}

\item{p_threshold}{a p_value indicating rejection threshold of Shapiro wilks test}
}
\value{
a float of the p-value and a string with the word Pass or Fail depending on the results of the test
}
\description{
This function receives a linear regression model and p-value threshold
and outputs the p-value from a Shapiro wilks test along with a statement
indicating the results of the normality test
}
\examples{
library(tibble)
normality(tibble("X"=c(2,3,4), "y"=c(2,5,6)), "y ~ X")

}
