% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicollinearity.R
\name{multicollinearity}
\alias{multicollinearity}
\title{multicollinearity}
\usage{
multicollinearity(data, formula, vif_threshold = 10)
}
\arguments{
\item{data}{a dataframe containing predictor data}

\item{formula}{a formula in the format "y ~ x1 + x2 + ..." indicating regression variables}

\item{threshold}{a float indicating threshold for VIF}
}
\value{
a dataframe containing VIF coefficients and a statement whether the assumption is violated.
}
\description{
This function receives aan input dat, linear regression model and vif threshold, and outputs VIF for each feature
along with a statement whether or not the multicollinearity assumption is violated.
}
\examples{
library(tibble)
data <- tibble(X = c(1,2,3,4,5))
formula <- "y ~ X"
multicollinearity(data, formula, 10)

}
